+++ inc/form_element_array.php
@@ -36,6 +36,10 @@ class form_element_array extends form_element {
 	$data[$k]=$element->get_data();
     }
 
+    if(!sizeof($data))
+      return array_key_exists('empty_value', $this->def) ?
+        $this->def['empty_value'] : null;
+
     return $data;
   }